% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ans.R
\name{fetch_ans}
\alias{fetch_ans}
\title{Fetch data from the Brazilian National Agency of Supplementary Health}
\usage{
fetch_ans(states, years)
}
\arguments{
\item{states}{character. One UF acronym or a vector.}

\item{years}{numeric. One year or a vector.}
}
\value{
A tibble.
}
\description{
Fetch data from the Brazilian National Agency of Supplementary Health
}
